Remove `WidgetViewModel`
