+++ PomeloCardsDemo/UI/ViewControllers/HomeViewController.swift
@@ -9,8 +9,8 @@ class HomeViewController: UIViewController {
     
     @IBOutlet weak var tableView: UITableView!
     
-    private var viewModel: WidgetViewModelProtocol = WidgetViewModel()
-
+    private let widgetsFactory = WidgetViewControllerFactory()
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         setupDelegates()
@@ -30,23 +30,51 @@ class HomeViewController: UIViewController {
 
 extension HomeViewController: UITableViewDelegate, UITableViewDataSource {
     
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int { viewModel.widgetsCount() }
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int { WidgetType.count }
     
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        guard let widgetType = WidgetType(rawValue: indexPath.row) else { return }
-        viewModel.presentWidget(widgetType, on: self)
+        guard let widget = WidgetType(rawValue: indexPath.row),
+              let widgetViewController = widgetsFactory.buildWidgetController(for: widget, params: widget.getParams()) else {
+            print("Couldn't select widget at indexPath: \(indexPath)")
+            return
+        }
+        if widget == .cardDetail {
+            self.displayViewControllerAsBottomSheet(widgetViewController)
+        } else {
+            self.present(widgetViewController, animated: true)
+        }
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         
         guard let cell = tableView.dequeueReusableCell(withIdentifier: WidgetTableViewCell.identifier,
                                                        for: indexPath) as? WidgetTableViewCell,
-              let widgetType = WidgetType(rawValue: indexPath.row) else {
+              let widget = WidgetType(rawValue: indexPath.row) else {
             print("Couldn't generate cell at indexPath: \(indexPath)")
             return WidgetTableViewCell()
         }
-        viewModel.configCell(cell, widget: widgetType)
+        cell.titleLabel.text = widget.getTitle()
         return cell
     }
 }
 
+extension WidgetType {
+    func getTitle() -> String {
+        switch self {
+        case .cardActivation: return "Activación de tarjeta"
+        case .changePin: return "Cambio de pin"
+        case .card: return "Tarjeta"
+        case .cardDetail: return "Datos de tarjeta"
+        }
+    }
+    
+    func getParams() -> [String: Any] {
+        switch self {
+        case .cardActivation: return [:]
+        case .changePin: return ["card_id": "crd-2H0AxFMF5XJGQqc6iSpUzMUS7Z3"]
+        case .card: return ["card_id": "crd-2H0AxFMF5XJGQqc6iSpUzMUS7Z3"]
+        case .cardDetail: return ["card_id": "crd-2H0AxFMF5XJGQqc6iSpUzMUS7Z3"]
+        }
+    }
+}
+
